Update to a more modern CMake using target properties (#558)
This makes a few changes moving to a more modern style. It adds our CMake module path to the CMake module path, it declares an empty target, adding source files to it instead of accumulating in variables. Also switched to add compiler definitions to the target, and avoid repetition for definitions used in C++ and CUDA.