Build binary pip wheels for cupynumeric (#610)
Build a Python pip binary wheel for the project. One odd thing to note is the `cupynumeric/_version.py` file needs to be moved out of the way for pip wheel builds. I would like to move conda builds to scikit-build-core, but keep this hack temporarily for wheels builds. There is also logic to avoid building for the CPU on the system in `cupynumeric` and `openblas`.

The wheel build fetches a wheel from the `legate` action, and that also needs fixing once we have pip wheels getting published. The tests are being run for CPU and CUDA on the different architectures we build for.