Migrate to a `gh` CLI based artifact download (#661)
Add some utility scripts to use the official `gh` CLI to look up the run ID and download artifacts. This enables us to migrate away from the third-party action and offers a base of capabilities as we split the conda packages and pip wheels.

This is inspired by and shamelessly takes from https://github.com/rapidsai/gha-tools/, specifically `rapids-download-from-github` and `rapids-github-run-id` adapting them for these projects.