Use a clean version before the build starts (#667)
The version file is ironically causing the version string to be incorrect. Issue #666 documents what should be done to fix this correctly. In the short term I need a clean release version for the upcoming initial wheels release hence overriding it.

Longer term we should override in CI anyway on the release branch to switch to a post release versioning scheme in release branches.