Improve legate wheel download error checking (#967)
Improve error checking if the run ID is not found for the legate wheel from the SHA. Add some improvements to the search query as well. This was failing in a surprising way where no run found resulted in an empty string which the `gh` CLI didn't error on, but instead consumed all API calls for the user over time.